+++ psmrts/shapes/obj/tests/test_PsmrtsOBJFormat.cpp
@@ -56,6 +56,19 @@ TEST_CASE ( "OBJ FORMAT Asset Test - Basic Load/Init Tests", "[format][obj][benn
     options3["required"] = {"obj_file"};
     options3["optional"] = {"obj_data_type", "obj_mtl_search_path"};
     psmrts::ProductSpecification test3("obj", "mesh", options2);
+
+    tinyobj::ObjReader reader;
+    REQUIRE( reader.ParseFromFile(objfile) == true );
+    // Is param intentional? shape() returns const, function accepts non-const version
+    psmrts::ProductConfiguration meta_data = t_loader.get_metadata( &reader ); 
+    CHECK( meta_data.name() == "" );
+    CHECK( meta_data.size() == 6 );
+    CHECK( psmrts::psmrts_filename(meta_data.find("obj_file").to_string()) == "bennu_20facets.obj" );
+    CHECK( meta_data.find("obj_data_type").to_string()       == "double" );
+    CHECK( meta_data.find("obj_mtl_search_path").to_string() == "" );
+    CHECK( meta_data.find("obj_shapes").to_string()          == "1" );
+    CHECK( meta_data.find("obj_vertices").to_string()        == "20" );
+    CHECK( meta_data.find("obj_facets").to_string()          == "36" );
 }
 
 TEST_CASE ( "OBJ FORMAT Asset Test - Load Fail Tests", "[format][obj][failures]" ) {