+++ psmrts/shapes/tests/test_PsmrtsShape.cpp
@@ -20,6 +20,9 @@ TEST_CASE("PsmrtsShape Default Test", "[shape][default]") {
     psmrts::PsmrtsMeshData test_data = test_shape.get_mesh();
     CHECK( test_data.nvectors() == 0 );
     CHECK( test_data.nfacets()  == 0 );
+
+    // This test is not the same on all OSes but tracks the size PSMRTS shapes!
+    // CHECK( sizeof( psmrts::PsmrtsShape ) == 1288 );
 }
 
 TEST_CASE("NAIF DSK Shape Test", "[naif][shape]") {