+++ CHANGELOG.md
@@ -35,6 +35,12 @@ release.
 ## TODO
 - See [TODO](TODO.md)
 
+## [0.3.7] 2025-11-04
+- Additions/updates to product configurations and tracking of products.
+- Update tracers to use private implementation, add construction options, tracking/meta data reporting and tests.
+- Add/update numerous tests.
+- Add Doxygen documentation to CMake system.
+
 ## [0.3.6] 2025-09-15
 - Removed unnecessary #include's
 - Fixed small number of minor misspellings.